Issue #3281 - Move template filters into template module
